+++ breeze-spring-boot/src/main/kotlin/DataAutoConfiguration.kt
@@ -1,9 +1,9 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
-package com.windea.breezeframework.springboot.data
+package com.windea.breezeframework.springboot
 
-import com.windea.breezeframework.springboot.data.converters.*
+import com.windea.breezeframework.springboot.converter.*
 import org.springframework.boot.autoconfigure.condition.*
 import org.springframework.context.annotation.*
 import org.springframework.data.domain.*