+++ breeze-spring-boot/src/main/kotlin/converter/StringToPageableConverter.kt
@@ -1,9 +1,8 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
-package com.windea.breezeframework.springboot.data.converters
+package com.windea.breezeframework.springboot.converter
 
-import com.windea.breezeframework.core.extensions.*
 import org.springframework.core.convert.converter.*
 import org.springframework.data.domain.*